[android] Migrate remaining uses of jint, jlong and friends
I used the script below as a basis, then did some manual adjustments,
particularly in  base/android/jni_array* (needs investigation).

```
git grep -El 'j(int|long|byte|char|short)' '*.h' '*.cc' |
  xargs -r sed -i -E -e 's/jint(\W)/int32_t\1/g' \
                     -e 's/jlong(\W)/int64_t\1/g' \
                     -e 's/jbyte(\W)/int8_t\1/g' \
                     -e 's/jchar(\W)/uint16_t\1/g' \
                     -e 's/jshort(\W)/int16_t\1/g'
git diff --pretty="" --name-only  |
  xargs -r tools/add_header.py --header '<cstdint>'
git grep -El 'j(boolean|float|double)' '*.h' '*.cc' |
  xargs -r sed -i -E -e 's/jboolean(\W)/bool\1/g' \
                     -e 's/jfloat(\W)/float\1/g' \
                     -e 's/jdouble(\W)/double\1/g'
git checkout third_party
```

Bug: 331706917
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7531657
Owners-Override: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/main@{#1578152}

+++ components/sync/android/sync_service_android_bridge.cc
@@ -5,6 +5,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "components/sync/android/sync_service_android_bridge.h"
 
+#include <cstdint>
 #include <map>
 #include <string>
 #include <vector>
@@ -181,14 +182,14 @@ void SyncServiceAndroidBridge::OnSyncShutdown(SyncService* sync) {
 
 void SyncServiceAndroidBridge::AcknowledgeBookmarksLimitExceededError(
     JNIEnv* env,
-    jint source) {
+    int32_t source) {
   native_sync_service_->AcknowledgeBookmarksLimitExceededError(
       static_cast<SyncService::BookmarksLimitExceededHelpClickedSource>(
           source));
 }
 
-jint SyncServiceAndroidBridge::GetBookmarksLimit(JNIEnv* env) {
-  return static_cast<jint>(kSyncBookmarksLimitValue.Get());
+int32_t SyncServiceAndroidBridge::GetBookmarksLimit(JNIEnv* env) {
+  return kSyncBookmarksLimitValue.Get();
 }
 
 bool SyncServiceAndroidBridge::IsSyncFeatureEnabled(JNIEnv* env) {
